Move attachment and subattachment dispatching to _file.py
